refactor: remove legacy config code
We don't support post platform config hooks at the moment.
These need to be reimplemented in "collie foundation new".
Also, e.g. the BigQuery table setup can be better served from
"collie foundation deploy" and a kit module rather than deploying
the bigquery dataset via the bq command line
